hardware-bridges/enderscope/requirements.txt
pyserial>=3.5
flask>=2.0
flask-cors>=3.0
Pillow>=9.0
